+++ pkg/provider/models.go
@@ -277,6 +277,30 @@ type CustomMetricJobResourceModel struct {
 	IsModelSpecific        types.Bool    `tfsdk:"is_model_specific"`
 }
 
+type CustomMetricFromJobResourceModel struct {
+	ID                 types.String             `tfsdk:"id"`
+	DeploymentID       types.String             `tfsdk:"deployment_id"`
+	CustomJobID        types.String             `tfsdk:"custom_job_id"`
+	Name               types.String             `tfsdk:"name"`
+	Description        types.String             `tfsdk:"description"`
+	BaselineValue      types.Float64            `tfsdk:"baseline_value"`
+	Timestamp          *MetricTimestampSpoofing `tfsdk:"timestamp"`
+	Value              *ColumnNameValue         `tfsdk:"value"`
+	Batch              *ColumnNameValue         `tfsdk:"batch"`
+	SampleCount        *ColumnNameValue         `tfsdk:"sample_count"`
+	Schedule           *Schedule                `tfsdk:"schedule"`
+	ParameterOverrides types.List               `tfsdk:"parameter_overrides"`
+}
+
+type MetricTimestampSpoofing struct {
+	ColumnName types.String `tfsdk:"column_name"`
+	TimeFormat types.String `tfsdk:"time_format"`
+}
+
+type ColumnNameValue struct {
+	ColumnName types.String `tfsdk:"column_name"`
+}
+
 // RegisteredModelResourceModel describes the registered model resource.
 type RegisteredModelResourceModel struct {
 	ID                   types.String   `tfsdk:"id"`